% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter_matrix.R
\name{adapter_matrix}
\alias{adapter_matrix}
\title{Create adapter matrix}
\usage{
adapter_matrix(adapters, ids = NULL, value = "pident", cores = 1)
}
\arguments{
\item{adapters}{adapter; a data frame of class "adapter" containing
identified adapter sequences.}

\item{ids}{character; ids within the 'adapters' data frame, in the same
order as they should appear in the matrix. If NULL (default), order will be
determined by the order of unique ids in the 'adapters' data frame. If
specified, all ids must be present in the'adapters' data frame and all ids
in the 'adapters' data frame must be present in the 'ids' vector.}

\item{value}{character; the variable within the adapter data frame to use as
values in the matrix, "pident", "mean_score", or "end".}

\item{cores}{integer; number of CPU cores to use.}
}
\description{
Convert identified adapters into matrix form.
}
\examples{
# import example data
data(rbps)
# import adapters identified from example data
data(adapters)
# convert to matrix and plot
adapters |> adapter_matrix() |> plot()
# use same order as in the original data frame
adapters |> adapter_matrix(ids = rbps$Core_ORF) |> plot()
}
