% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor-package.R
\docType{data}
\name{adapters}
\alias{adapters}
\title{Adapters}
\format{
A data frame with 90 rows and 6 variables:
\describe{
  \item{pattern_id}{pattern_id}
  \item{subject_id}{subject_id}
  \item{start}{location of the first amino acid within the adapter sequence}
  \item{end}{location of the last amino acid within the adapter sequence}
  \item{mean_score}{mean amino acid position score between 'start' and 'end'}
  \item{pident}{ratio of identical amino acids between 'start' and 'end'}
}
}
\usage{
adapters
}
\description{
This data set contains identified adapters from all pairwise comparisons of 
the RBPs in the `rbps` data set. The approach to calculate these adapters
was the following: first, check for adapters with the "xbar.one" method, 
because it seems to be highly specific (i.e. it rarely identifies adapters 
when there isn't one). If an adapter was found, then the "cemean" method was 
used to get a more accurate estimate of the adapter length.
}
\keyword{datasets}
