% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-equivocal-zone.R
\name{adjust_equivocal_zone}
\alias{adjust_equivocal_zone}
\title{Apply an equivocal zone to a binary classification model.}
\usage{
adjust_equivocal_zone(x, value = 0.1, threshold = 1/2)
}
\arguments{
\item{x}{A \code{\link[=container]{container()}}.}

\item{value}{A numeric value (between zero and 1/2) or \code{\link[hardhat:tune]{hardhat::tune()}}. The
value is the size of the buffer around the threshold.}

\item{threshold}{A numeric value (between zero and one) or \code{\link[hardhat:tune]{hardhat::tune()}}.}
}
\description{
Apply an equivocal zone to a binary classification model.
}
\examples{
library(dplyr)
library(modeldata)

post_obj <-
  container(mode = "classification") \%>\%
  adjust_equivocal_zone(value = 1 / 4)


post_res <- fit(
  post_obj,
  two_class_example,
  outcome = c(truth),
  estimate = c(predicted),
  probabilities = c(Class1, Class2)
)

predict(post_res, two_class_example)
}
