% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-numeric-range.R
\name{adjust_numeric_range}
\alias{adjust_numeric_range}
\title{Truncate the range of numeric predictions}
\usage{
adjust_numeric_range(x, lower_limit = -Inf, upper_limit = Inf)
}
\arguments{
\item{x}{A \code{\link[=tailor]{tailor()}}.}

\item{upper_limit, lower_limit}{A numeric value, NA (for no truncation) or
\code{\link[hardhat:tune]{hardhat::tune()}}.}
}
\description{
Truncate the range of numeric predictions
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(tibble)

# create example data
set.seed(1)
d <- tibble(y = rnorm(100), y_pred = y/2 + rnorm(100))
d

# specify calibration
tlr <-
  tailor() \%>\%
  adjust_numeric_range(lower_limit = 1)

# train tailor by passing column names. situate in a modeling workflow with
# `workflows::add_tailor()` to avoid having to specify column names manually
tlr_fit <- fit(tlr, d, outcome = y, estimate = y_pred)

predict(tlr_fit, d)
\dontshow{\}) # examplesIf}
}
