% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-predictions-custom.R
\name{adjust_predictions_custom}
\alias{adjust_predictions_custom}
\title{Change or add variables}
\usage{
adjust_predictions_custom(x, ..., .pkgs = character(0))
}
\arguments{
\item{x}{A \code{\link[=tailor]{tailor()}}.}

\item{...}{Name-value pairs of expressions. See \code{\link[dplyr:mutate]{dplyr::mutate()}}.}

\item{.pkgs}{A character string of extra packages that are needed to execute
the commands.}
}
\description{
This adjustment functions allows for arbitrary transformations of model
predictions using \code{\link[dplyr:mutate]{dplyr::mutate()}} statements.
}
\examples{
\dontshow{if (rlang::is_installed("modeldata")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(modeldata)

head(two_class_example)

tlr <-
  tailor() \%>\%
  adjust_equivocal_zone() \%>\%
  adjust_predictions_custom(linear_predictor = binomial()$linkfun(Class2))

tlr_fit <- fit(
  tlr,
  two_class_example,
  outcome = c(truth),
  estimate = c(predicted),
  probabilities = c(Class1, Class2)
)

predict(tlr_fit, two_class_example) \%>\% head()
\dontshow{\}) # examplesIf}
}
