% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-probability-calibration.R
\name{adjust_probability_calibration}
\alias{adjust_probability_calibration}
\title{Re-calibrate classification probability predictions}
\usage{
adjust_probability_calibration(x, method = NULL)
}
\arguments{
\item{x}{A \code{\link[=container]{container()}}.}

\item{method}{Character. One of \code{"logistic"}, \code{"multinomial"},
\code{"beta"}, \code{"isotonic"}, or \code{"isotonic_boot"}, corresponding to the
function from the \pkg{probably} package \code{\link[probably:cal_estimate_logistic]{probably::cal_estimate_logistic()}},
\code{\link[probably:cal_estimate_multinomial]{probably::cal_estimate_multinomial()}}, etc., respectively.}
}
\description{
Re-calibrate classification probability predictions
}
