% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_GMM_utils.R
\name{bulk_weighted_gmm}
\alias{bulk_weighted_gmm}
\title{bulk_weighted_gmm}
\usage{
bulk_weighted_gmm(
  data,
  k,
  params,
  weights = NULL,
  initialize = NULL,
  seed = 91,
  steps = 10,
  regularize_variance = TRUE,
  variance_correction = NULL,
  relative_loglik_threshold = 0.005,
  verbose = FALSE
)
}
\arguments{
\item{data}{a matrix containing events along the rows, variables along columns}

\item{k}{the number of mixture components (clusters)}

\item{weights}{a vector of length = nrow(data), specifying the weight of each event}
}
\value{
A gmm object
}
\description{
This function calculates a bulk weighted gmm
}
