% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_adapters.R
\name{cluster_adapters}
\alias{cluster_adapters}
\title{Group RBPs into adapter clusters}
\usage{
cluster_adapters(
  mat,
  k_min = 1,
  k_max = NULL,
  h = NULL,
  homogeneity_thr = 0.75,
  completeness_thr = 0.75,
  cores = 1,
  verbose = getOption("verbose")
)
}
\arguments{
\item{mat}{adapter_matrix; a matrix of class "adapter_matrix". Use 
adapter_matrix() to create a compatible matrix.}

\item{k_min}{integer; minimum number of clusters to split the RBPs into.}

\item{k_max}{integer; maximum number of clusters to split the RBPs into.}

\item{h}{hclust object; a hierarchical clustering object created from the 
distance matrix. If NULL, the function will create a hierarchical clustering 
object from the input matrix using the average method.}

\item{homogeneity_thr}{numeric; minimum percentage identity of the identified
adapter sequence to be considered a shared adapter for homogeneity.}

\item{completeness_thr}{numeric; minimum percentage identity of the
identified adapter sequence to be considered a shared adapter for
completeness.}

\item{cores}{integer; number of CPU cores to use.}

\item{verbose}{logical; should verbose messages be printed to the console?
If TRUE, a progress bar is also displayed.}
}
\value{
a data frame with two columns, id and cluster, containing cluster
assignments.
}
\description{
Group RBPs into adapter clusters
}
\examples{
data(rbps)
data(adapters)
# convert to pident matrix and order by the original data frame
amat <- adapter_matrix(adapters, ids = rbps$Core_ORF)
# cluster RBPs into 2 to 5 clusters
cluster_adapters(amat, k_min = 2, k_max = 5)
}
