% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container-package.R
\docType{package}
\name{container-package}
\alias{container-package}
\title{container: Sandbox for a postprocessor object}
\description{
Sandbox for a postprocessor object.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/container}
  \item Report bugs at \url{https://github.com/tidymodels/container/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co}

Authors:
\itemize{
  \item Simon Couch \email{simon.couch@posit.co}
  \item Hannah Frick \email{hannah@posit.co}
  \item Emil HvitFeldt \email{emil.hvitfeldt@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
