% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{container}
\alias{container}
\title{Declare post-processing for model predictions}
\usage{
container(
  mode = "unknown",
  type = "unknown",
  outcome = character(0),
  estimate = character(0),
  probabilities = character(0),
  time = character(0),
  call = current_env()
)
}
\arguments{
\item{mode}{The model's mode, one of \code{"unknown"}, \code{"classification"}, or
\code{"regression"}. Modes of \code{"censored regression"} are not currently supported.}

\item{type}{The model sub-type. Possible values are \code{"unknown"}, \code{"regression"},
\code{"binary"}, or \code{"multiclass"}.}

\item{outcome}{The name of the outcome variable.}

\item{estimate}{The name of the point estimate (e.g. predicted class)}

\item{probabilities}{The names of class probability estimates (if any). For
classification, these should be given in the order of the factor levels of
the \code{estimate}.}

\item{time}{The name of the predicted event time. (not yet supported)}

\item{call}{The call to be displayed in warnings or errors.}
}
\description{
Declare post-processing for model predictions
}
\examples{

container()
}
