% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_comparisons.R
\name{filter_comparisons}
\alias{filter_comparisons}
\title{Filter comparisons}
\usage{
filter_comparisons(comps, filter_value, filter_by = "Core_ORF", data = NULL)
}
\arguments{
\item{comps}{data.frame; the comparisons data set to be filtered. Must be of
class "breakpoints" or "adapter". See \code{\link{find_breakpoints}},
\code{\link{find_adapter}} or \code{\link{find_all_adapters}} for compatible
data frames.}

\item{filter_value}{character; the value to filter by. By default, this is
a pattern_id or subject id from the comparisons data frame, but can be any
value in the supporting data set if filter_by and data are specified.}

\item{filter_by}{character; name of the variable in the supporting data set
to filter by. By default, this is "Core_ORF", but can be any column in the
supporting data set.}

\item{data}{data.frame; the supporting data set that links ids with
properties. This is required if filter_by is not "Core_ORF". Must contain a
column with the same ids as the comparisons data frame (e.g. "Core_ORF") and
a column with the values to filter by (e.g. "ACL 2").}
}
\description{
Filter comparisons (brakpoints or adapters) based on ids or properties from
the supporting data set.
}
\examples{
data(rbps)
data(adapters)
# convert to pident matrix and order by the original data frame
amat <- adapter_matrix(adapters, ids = rbps$Core_ORF)
# cluster RBPs into 2 to 5 clusters
clusters <- cluster_adapters(amat, k_min = 2, k_max = 5)
rbps <- dplyr::left_join(rbps, clusters, by = c("Core_ORF" = "id"))
# adapters with "ON513429-1"
adapters |> filter_comparisons("ON513429-1")
# adapter between "MN395291-1" and "ON513429-1"
adapters |> filter_comparisons("MN395291-1") |> filter_comparisons("ON513429-1")
# adapters in ACL 2
adapters |> filter_comparisons("ACL 2", filter_by = "cluster", data = rbps)
}
