% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_all_adapters.R
\name{find_all_adapters}
\alias{find_all_adapters}
\title{Find all adapters for a set of RBPs}
\usage{
find_all_adapters(
  ids,
  data,
  id_var = "Core_ORF",
  seq_var = "translation",
  submat = "BLOSUM80",
  max_end_vars = NULL,
  max_end = NULL,
  method = "cemean",
  min_pident = 0.4,
  max_start = 10,
  merge_beginning = TRUE,
  merge_conserved = TRUE,
  cores = 1,
  verbose = getOption("verbose"),
  ...
)
}
\arguments{
\item{ids}{character; the names of RBPs to compare.}

\item{data}{data.frame; a data frame that contains IDs and sequences}

\item{id_var}{character; variable within \code{data} that stores IDs.}

\item{seq_var}{character; variable within \code{data} that stores sequences.}

\item{submat}{character; the substitution matrix. See
\code{position_scores()} for more information.}

\item{max_end_vars}{character; optional vector of variable names in
\code{data} that contain numeric values representing positions which the
adapters should not overlap with and terminate earlier (e.g., domain
start positions). See Details for more information.}

\item{max_end}{integer; the maximum position to look for breakpoints. If 
`NULL`, search the full alignment, if an integer, limit the search to that 
position. If both this argument and `position_scores$max_end` are set, the 
smaller integer is used.}

\item{method}{character; the method to use for finding breakpoints. Either
"cemean", "plateau", or "window". See \code{find_breakpoints()} for more
information}

\item{min_pident}{numeric; the lowest accepted percentage identity for a
region to be considered a "conserved" region.}

\item{max_start}{integer; the maximum accepted starting position for the
first conserved region.}

\item{merge_beginning}{logical; whether to merge the positions before the
first conserved region. See Details for more information.}

\item{merge_conserved}{logical; whether neighboring conserved regions should
be merged. See Details for more information.}

\item{cores}{integer; number of CPU cores to use.}

\item{verbose}{logical; should verbose messages be printed to the console?}

\item{...}{additional arguments to pass to breakpoint detection. See
\code{find_breakpoints()} for more information.}
}
\value{
A data.frame of class "adapter".
}
\description{
Look at all pairs between a set of RBPs, align, detect breakpoints, find
adapters and combine them into a single data frame.
}
\details{
If you provide one or more `max_end_var` names, the function will
look for their values in both the pattern and subject rows, calculate their
minimum  and return it in the output object. The value will then be
respected by `find_breakpoints()` and the function will not look at
positions beyond this value.
}
\examples{
data(rbps)
find_all_adapters(rbps$Core_ORF[1:3], data = rbps)
}
