% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_breakpoints.R
\name{find_breakpoints}
\alias{find_breakpoints}
\alias{find_breakpoints_cemean}
\alias{find_breakpoints_plateau}
\alias{find_breakpoints_window}
\title{Find breakpoints in global alignment scores}
\usage{
find_breakpoints(position_scores, method = "cemean", ...)

find_breakpoints_cemean(position_scores, Nmax = 5, seed = 0)

find_breakpoints_plateau(position_scores, pinit = 0.05, type = "ewma", ...)

find_breakpoints_window(
  position_scores,
  window = 5,
  score_threshold = 3,
  pident_threshold = 0.4,
  p_threshold = 0.05
)
}
\arguments{
\item{position_scores}{an object of class `ps`, returned by
`position_scores()`.}

\item{method}{character; the method to use for finding breakpoints. Either
"cemean", "plateau", or "window".}

\item{...}{additional arguments depending on the selected method. When
`method = "plateau"`, additional arguments passed on to the selected control
chart. See `?qcc::cusum()`, `?qcc::ewma()`, or `?qcc::qcc()` for a full list
of control chart parameters.}

\item{Nmax}{integer; the maximum number of breakpoints to look for.}

\item{seed}{integer; random seed.}

\item{pinit}{numeric; ratio of sequence length used for setting the center.
See Details for more information.}

\item{type}{character; the type of statistical process control chart to use.
Currently supported cards are `cusum`, `ewma` and `xbar.one`.}

\item{window}{integer; the number of successive amino acids to look at}

\item{score_threshold}{numeric; mu threshold for alignment scores}

\item{pident_threshold}{numeric; mu threshold for amino acid identities}

\item{p_threshold}{numeric; p value threshold of significance}
}
\value{
A data frame with 6 columns:
\itemize{
  \item `pattern_id`: pattern ID
  \item `subject_id`: subject ID
  \item `start`: first position for a section of the alignment
  \item `end`: last position for a section of the alignment
  \item `mean_score`: mean score for amino acid pairs within the section
  \item `pident`: ratio of matching amino acid pairs within the section
}
}
\description{
This function looks at global alignment scores by position and attempts to
find locations along the length of the aligned sequences where these scores
tend to change substantially.
}
\details{
When using `method = "cemean"` the function will estimate the number
of breakpoints and their locations using the Cross-Entropy Method. This
method will use the `CE.Normal.Mean()` function from the `breakpoint` package
to find the breakpoints.

When using `method = "plateau"` the function will look at the first
few positions of the alignment (goverened by `pinit`) and calculate a mean
score. This mean score will be used as the center: 1. it will be used as the
center point for a number of control charts; 2. any scores above the center
will be reduced to be indentical with the center. Then the function will use
the control chart selected by `type`, and extract sections using the
violations.

When using `method == "window"` the function will form disjunct
windows along the length of the alignment and look at the positions within
each window. For each window, it will perform one sided wilcoxon tests
against the `score_threshold` and the `pident_threshold` and consider the
window conserved if at least one of the tests passes. Finally it will
collapse neighboring windows that are all considered conserved or
non-conserved and return a data frame of sections in which conservation
alternates between the lines.
}
\note{
When using `method == "cemean"` the speed of the function will depend
on the number of breakpoints to look for.
}
\examples{
data(rbps)
ps <- position_scores("MN395291-1", "ON513429-1", data = rbps)

# Find breakpoints using the Cross-Entropy Method
find_breakpoints(ps, method = "cemean", Nmax = 5)
# Find breakpoints using plateau on scores and EWMA chart
find_breakpoints(ps, method = "plateau", type = "ewma", lambda = 0.2)
# Find breakpoints using plateau on scores and CUSUM chart
find_breakpoints(ps, method = "plateau", type = "cusum")
# Find breakpoints using disjunct windows
find_breakpoints(ps, method = "window", window = 5)
}
\seealso{
[position_scores()]
}
