% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_methods.R
\name{inspect_1D_mixtures}
\alias{inspect_1D_mixtures}
\title{inspect_1D_mixtures}
\usage{
inspect_1D_mixtures(data, mixtures_1D, params)
}
\arguments{
\item{data}{A flowFrame or a matrix containing events along the rows, markers along columns.}

\item{mixtures_1D}{1D mixture models, as produced by get_1D_mixtures.}

\item{params}{A list of markers to use; must be subset of colnames(data).}
}
\description{
Plot the result of 1D mixture model calculation for visual inspection.
Displays, for each marker, three side-by-side plots, giving a kernel density estimate
for the data and that marker, the Gaussian mixture, and the separate mixture components,
respectively.
}
