% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_methods.R
\name{make_cluster_phenobars}
\alias{make_cluster_phenobars}
\title{make_cluster_phenobars}
\usage{
make_cluster_phenobars(tailor_obj, cluster_ids)
}
\arguments{
\item{tailor_obj}{A tailor object, as obtained from tailor.learn.}

\item{cluster_ids}{A list of cluster indices to visualize. For example, those deemed
differentially expressed in case/controls by some statistical test.}
}
\value{
Boxplots summarizing means and standard deviations for selected clusters.
}
\description{
Visualize summary of cluster information, via boxplots showing means
and standard deviations of mean fluorescence intensity. Only markers with at least
one cutoff are displayed. The boxes are color-coded by mfi relative to the cutoff.
}
\examples{
fileName <- system.file("extdata", "sampled_flowset_old.rda",
                       package = "Tailor")
load(fileName)
tailor_params <- flowCore::colnames(fs_old)[c(7:9, 11:22)]
tailor_obj <- tailor_learn(data = fs_old,
                          params = tailor_params,
                          mixture_components = 50)

make_cluster_phenobars(tailor_obj, cluster_ids = c(1,2))
}
