% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_adapter_matrix.R
\name{plot.adapter_matrix}
\alias{plot.adapter_matrix}
\title{Plot adapter matrix}
\usage{
\method{plot}{adapter_matrix}(x, clusters = NULL, ...)
}
\arguments{
\item{x}{adapter_matrix; a matrix of class "adapter_matrix". Use 
adapter_matrix() to create a compatible matrix.}

\item{clusters}{data.frame; a table of RBP IDs and cluster designations.}

\item{...}{additional arguments (not used).}
}
\value{
A heatmap. If clusters are provided, also show cluster assignments.
}
\description{
Take an adapter matrix and optionally data frame of cluster designations and
plot them on a simple heatmap.
}
\examples{
data(rbps)
data(adapters)
# convert to pident matrix and order by the original data frame
amat <- adapter_matrix(adapters, ids = rbps$Core_ORF)
# plot pident matrix
plot(amat)
# split RBPs into two clusters
clusters <- cluster_adapters(amat, k_min = 2, k_max = 2)
# plot pident matrix with cluster assignments
plot(amat, clusters = clusters)
}
