% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_position_scores.R
\name{plot.ps}
\alias{plot.ps}
\title{Plot pairwise alignment scores with adapter regions}
\usage{
\method{plot}{ps}(x, type = "ma", method = NULL, highlight = NULL, size = 10, ...)
}
\arguments{
\item{x}{an object of class `ps`, returned by `position_scores()`.}

\item{type}{character; strategy for plotting data points. Either "indiv",
"ma", or "cusum". See Details for more information.}

\item{method}{character; the method to use for finding the conserved N
terminal region. Can be one or more of "cemean", "cusum", "ewma", "xbarone",
and "window".}

\item{highlight}{character; the method to use for highlighting the conserved
N terminal region. Must be one of the methods used in `method`.}

\item{size}{integer; the number of successive data point to use for
calculating moving averages. Only used if `type == "ma"`.}

\item{...}{additional arguments (not used).}
}
\value{
a plot.
}
\description{
This is a convenience function for looking at pairwise alignments. The
function will plot alignment scores along the length of the alignment. At the
same time, the function will also attempt to find an adapter region using 
various approaches and add the end of the conserved region to the plot.
}
\details{
When `type = "indiv"` each alignment score will be plotted
separately. When `type = "ma"` the plot will show moving averages. When
`type = "cusum"` the plot will show the cumulative sum of scores by position.
}
\examples{
data(rbps)
ps <- position_scores("MN395291-1", "ON513429-1", data = rbps)
plot(ps)
}
