% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_methods.R
\name{plot_kdes_global}
\alias{plot_kdes_global}
\title{plot_kdes_global}
\usage{
plot_kdes_global(data, params)
}
\arguments{
\item{data}{A flowSet, flowFrame or a matrix containing events along the rows, markers along columns.}

\item{params}{A list of markers to use; must be subset of colnames(data).}
}
\value{
Plots of kernel density estimate for each chosen parameter.
}
\description{
Plot 1D kdes of a dataset for visual inspection.
}
