% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_methods.R
\name{plot_tailor_fluorescence}
\alias{plot_tailor_fluorescence}
\title{plot_tailor_fluorescence}
\usage{
plot_tailor_fluorescence(tailor_obj)
}
\arguments{
\item{tailor_obj}{A tailor object, as obtained from tailor.learn.}
}
\value{
A grid of 2d reduced plots of cluster centroids, one for each
parameter.
}
\description{
Plot a t-SNE reduction to 2d of the cluster centroids,
color-coded by mean fluorescence intensity for each parameter.
}
\examples{
fileName <- system.file("extdata", "sampled_flowset_old.rda",
                       package = "Tailor")
load(fileName)
tailor_params <- flowCore::colnames(fs_old)[c(7:9, 11:22)]
tailor_obj <- tailor_learn(data = fs_old,
                          params = tailor_params,
                          mixture_components = 50)

plot_tailor_fluorescence(tailor_obj)
}
