% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_methods.R
\name{plot_tailor_majpheno}
\alias{plot_tailor_majpheno}
\title{plot_tailor_majpheno}
\usage{
plot_tailor_majpheno(tailor_obj)
}
\arguments{
\item{tailor_obj}{A tailor object, as obtained from tailor.learn.}
}
\value{
A 2d reduced plot of cluster centroid,
color-coded by major phenotype.
}
\description{
Plot a t-SNE reduction to 2d of the cluster centroids.
}
\examples{
fileName <- system.file("extdata", "sampled_flowset_old.rda",
                         package = "Tailor")
load(fileName)
tailor_params <- flowCore::colnames(fs_old)[c(7:9, 11:22)]
tailor_obj <- tailor_learn(data = fs_old,
                          params = tailor_params,
                          mixture_components = 50)

defs <- read.csv("~/git/R/packages/Tailor/inst/extdata/pheno_definitions.csv",
                 row.names = 1, stringsAsFactors = FALSE)
tailor_obj <- categorical_labeling(tailor_obj, defs)

plot_tailor_majpheno(tailor_obj)
}
