% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_methods.R
\name{plot_tsne_bin_centers}
\alias{plot_tsne_bin_centers}
\title{plot_tsne_bin_centers}
\usage{
plot_tsne_bin_centers(tailor_obj)
}
\arguments{
\item{tailor_obj}{A tailor object, as obtained from tailor.learn.}
}
\value{
A 2d reduced plot of bin centers, which can be seen as representatives for the cell
population. Color coded by major phenotype.
}
\description{
Plot a t-SNE reduction to 2d of the cluster centroids.
}
