% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_methods.R
\name{plot_tsne_global}
\alias{plot_tsne_global}
\title{plot_tsne_global}
\usage{
plot_tsne_global(data, tailor_obj, tailor_pred, defs)
}
\arguments{
\item{data}{A flowSet, flowFrame or a matrix containing events along the rows, markers along columns.}

\item{tailor_obj}{A tailor object, as obtained from tailor.learn.}

\item{tailor_pred}{A list of cluster assignments, as obtained from tailor.predict.}

\item{defs}{Definitions of major phenotypes.}
}
\value{
2d embeddings, color-coded by cluster membershi and major phenotype.
}
\description{
Subsample the entire dataset, and compute a 2D t-SNE embedding.
Plot the embedding twice, color coded by cluster membership and major phenotype, respectively.
}
