% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_scores.R
\name{position_scores}
\alias{position_scores}
\title{Pairwise global alignments with position scores}
\usage{
position_scores(
  pattern_id,
  subject_id,
  data,
  id_var = "Core_ORF",
  seq_var = "translation",
  submat = "BLOSUM80",
  constrain_vars = NULL,
  verbose = getOption("verbose")
)
}
\arguments{
\item{pattern_id}{character; pattern ID}

\item{subject_id}{character; subject ID}

\item{data}{data.frame; a data frame that contains IDs and sequences}

\item{id_var}{character; variable within \code{data} that stores IDs.}

\item{seq_var}{character; variable within \code{data} that stores sequences.}

\item{submat}{character; the substitution matrix. Options include
\code{"BLOSUM45"}, \code{"BLOSUM50"}, \code{"BLOSUM62"}, \code{"BLOSUM80"},
\code{"BLOSUM100"}, \code{"PAM30"}, \code{"PAM40"}, \code{"PAM70"},
\code{"PAM120"}, \code{"PAM250"}.}

\item{constrain_vars}{character; optional vector of variable names in 
\code{data} that contain numeric values representing positions which the 
adapters should not overlap with and terminate earlier (e.g., domain 
start positions). See Details for more information.}

\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
an object of class \code{ps} which is a list of four elements:
\code{pattern_id}, \code{subject_id}, \code{position_scores} and
\code{substitution_matrix}. \code{position scores} is a tibble that contains
nucleotide or amino acid identities and alignment scores by position.
}
\description{
This function takes a pair of IDs within a data set, performs a pairwise 
global alignment on their sequences, and returns an object which contains 
the pattern and subject IDs and a tibble of nucleotide or amino acid 
identities and alignment scores by position.
}
\details{
If you provide one or more constrain variables, the function will
look for these in both the pattern and subject rows, calculate their minimum 
and then filter the alignment to only include non overlapping positions.
}
\examples{
data(rbps)
position_scores("MN395291-1", "ON513429-1", data = rbps)
}
