% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor-package.R
\docType{data}
\name{rbps}
\alias{rbps}
\title{Receptor Binding Proteins}
\format{
A data frame with 20 rows and 2 variables:
\describe{
  \item{Core_ORF}{Unique identifier for each receptor binding protein}
  \item{translation}{Amino acid sequence of the protein}
}
}
\usage{
rbps
}
\description{
This data set contains example bacteriophage receptor binding proteins (RBPs) 
for demonstration purposes.
}
\keyword{datasets}
