% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoco.R
\name{schoco}
\alias{schoco}
\title{SCores of HOmogeneity and COmpleteness}
\usage{
schoco(mat, k, homogeneity_thr = 0.75, completeness_thr = 0.75)
}
\arguments{
\item{mat}{adapter_matrix; a matrix of class "adapter_matrix". Use
adapter_matrix() to create a compatible matrix.}

\item{k}{integer; the number of clusters to split the RBPs into.}

\item{homogeneity_thr}{numeric; minimum percentage identity of the identified
adapter sequence to be considered a shared adapter for homogeneity.}

\item{completeness_thr}{numeric; minimum percentage identity of the
identified adapter sequence to be considered a shared adapter for
completeness.}
}
\value{
A data frame with 6 columns: nclust (number of clusters requested),
cluster (cluser identifier), homogeneity (homogeneity score for the cluster),
completeness (completeness score for the cluster), schoco (sum of weighted
homogeneity and completeness scores, where homogeneity is multiplied by the
size of the cluster and completeness is multiplied by log2 of the size of the
cluster. The data frame contains one row for each cluster.
}
\description{
Calculate homogeneity and completeness scores for clusters of RBPs based on
the percentage identity of shared adapter sequences. The "schoco" score is a
weighted sum of the homogeneity and completeness scores, where homogeneity is
weighted by the size of the cluster and completeness is weighted by the log2
of the size of the cluster.
}
\examples{
data(rbps)
data(adapters)
# convert to pident matrix and order by the original data frame
amat <- adapter_matrix(adapters, ids = rbps$Core_ORF)
# split matrix into two clusters and calculate schoco values
schoco(amat, k = 2)
}
