% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tailor-internals}
\alias{tailor-internals}
\alias{tailor_fully_trained}
\alias{tailor_requires_fit}
\title{Internal tailor functions}
\usage{
tailor_fully_trained(x)

tailor_requires_fit(x)
}
\value{
\code{tailor_fully_trained()} and \code{tailor_requires_fit()} both
return a single logical for the condition described in the function names.
Each evaluates the condition for each adjustment method contained in the
\code{\link[=tailor]{tailor()}} object.
}
\description{
Utilities for use in downstream packages.
}
\keyword{internal}
