% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_methods.R
\name{tailor.learn}
\alias{tailor.learn}
\title{tailor.learn}
\usage{
tailor.learn(
  data,
  params = NULL,
  mixture_components = 200,
  min_bin_size = 20,
  max_bin_size = 200,
  mixtures_1D_path = NULL,
  seed = 135,
  do_tsne = TRUE,
  do_variance_correction = TRUE,
  parallel = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A matrix containing events along the rows, markers along columns.}

\item{params}{A list of markers to use; must be subset of colnames(data).}

\item{mixture_components}{The number of mixture components to learn. Some of these
are eventually merged, so it's a good idea to choose a number slightly larger than
the number of clusters you expect to get.}

\item{min_bin_size}{Bins with fewer events than this threshold are considered outliers, and are
ignored during the weighted EM algorithm. These events can still be assigned to clusters during
the prediction phase.}

\item{max_bin_size}{Bins with more events than this threshold are split, to ensure that the
weighted EM algorithm closely approximates a run of vanilla EM on the entire dataset.}

\item{mixtures_1D_path}{A path for pre-computed 1D mixture models, to be used for binning.
These are computed from scratch if the path is not provided.}

\item{seed}{A seed for the random number generator, to be used in the initialization of the bin
centers.}

\item{do_tsne}{Boolean flag; if true, computes t-SNE reduction to 2D of the bin centers,
for use in visualizations.}

\item{do_variance_correction}{Boolean flag; if true, corrects the variances of the mixture
components found in weighted EM, to account for intra-bin variance. TRUE is strongly recommended.}

\item{parallel}{Boolean flag; if true, uses multithreading to speed up computation.}

\item{verbose}{Boolean flag; if true, outputs timing and milestone information.}
}
\value{
A tailor object containing:
\describe{
\item{fit}{The tailor model, a named list containing the mixture proportions, means and variances
of all mixture components.}
\item{mixtures_1D}{}
\item{cat_clusters}{A named list containing information about the categorical clusters found by
the model: phenotype, cluster centers, and a mapping from mixture components to categorical clusters.}
\item{tsne_centers}{Optional: a dimensional reduction to 2D of the bin centers.}
}
}
\description{
This function learns a tailor model from input data. It
computes a preliminary binning of the data, then computes a mixture model using
a weighted version of the expectation-maximization (EM) algorithm,
and finally merges mixture components which are positive/negative for the same
markers, using adaptive thresholds.
}
