% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_methods.R
\name{tsne_clusters}
\alias{tsne_clusters}
\title{tsne_clusters}
\usage{
tsne_clusters(tailor_obj, seed = NULL)
}
\arguments{
\item{tailor_obj}{A tailor object, as obtained from tailor.learn.}

\item{seed}{A seed for the random generator, used in the initialization of t-SNE.}
}
\value{
A data frame containing the 2d embedding of the cluster centroids.
}
\description{
Obtain a t-SNE reduction to 2d of the cluster centroids.
}
