% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_rbps.R
\name{validate_rbps}
\alias{validate_rbps}
\title{Validate RBP data set}
\usage{
validate_rbps(
  rbps,
  id_var = "Core_ORF",
  seq_var = "translation",
  verbose = getOption("verbose")
)
}
\arguments{
\item{rbps}{data.frame}

\item{id_var}{name of the variable in \code{rbps} that contains unique
identifiers for each RBP. Must be unique and of class "character".}

\item{seq_var}{name of the variable in \code{rbps} that contains the amino
acid sequences. Must be of class "character" and cannot contain empty strings
or unknown characters.}

\item{verbose}{should the function return verbose messages to console?}
}
\value{
The function returns \code{TRUE} if the tails data set is consistent
and returns \code{FALSE} if it is not.
}
\description{
This function can be used to check the RBP data set for any consistency
issues that would produce errors in downstream analysis.
}
\details{
The function checks whether id_var is present in the data set and is
unique, and whether seq_var is present in the data set, is of class
"character", and does not contain empty strings. The function also checks
whether seq_var contains unknown characters, i.e. characters other than the
20 standard amino acids and the ambiguous characters B, Z, and X. If any of
these checks fail, the function returns \code{FALSE} and, if
\code{verbose = TRUE}, also returns a warning message to console.
}
\examples{
data(rbps)
validate_rbps(rbps, verbose = TRUE)
}
