% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-numeric-calibration.R
\name{adjust_numeric_calibration}
\alias{adjust_numeric_calibration}
\title{Re-calibrate numeric predictions}
\usage{
adjust_numeric_calibration(x, calibrator)
}
\arguments{
\item{x}{A \code{\link[=container]{container()}}.}

\item{calibrator}{A pre-trained calibration method from the \pkg{probably}
package, such as \code{\link[probably:cal_estimate_linear]{probably::cal_estimate_linear()}}.}
}
\description{
Re-calibrate numeric predictions
}
