% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-numeric-calibration.R
\name{adjust_numeric_calibration}
\alias{adjust_numeric_calibration}
\title{Re-calibrate numeric predictions}
\usage{
adjust_numeric_calibration(x, calibrator)
}
\arguments{
\item{x}{A \code{\link[=container]{container()}}.}

\item{calibrator}{A pre-trained calibration method from the \pkg{probably}
package, such as \code{\link[probably:cal_estimate_linear]{probably::cal_estimate_linear()}}.}
}
\description{
Re-calibrate numeric predictions
}
\examples{
library(modeldata)
library(probably)
library(tibble)

# create example data
set.seed(1)
dat <- tibble(y = rnorm(100), y_pred = y/2 + rnorm(100))

dat

# calibrate numeric predictions
reg_cal <- cal_estimate_linear(dat, truth = y, estimate = y_pred)

# specify calibration
reg_ctr <-
  container(mode = "regression") \%>\%
  adjust_numeric_calibration(reg_cal)

# "train" container
reg_ctr_trained <- fit(reg_ctr, dat, outcome = y, estimate = y_pred)

predict(reg_ctr, dat)
}
