% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-numeric-range.R
\name{adjust_numeric_range}
\alias{adjust_numeric_range}
\title{Truncate the range of numeric predictions}
\usage{
adjust_numeric_range(x, lower_limit = -Inf, upper_limit = Inf)
}
\arguments{
\item{x}{A \code{\link[=tailor]{tailor()}}.}

\item{upper_limit, lower_limit}{A numeric value, NA (for no truncation) or
\code{\link[hardhat:tune]{hardhat::tune()}}.}
}
\description{
Truncate the range of numeric predictions
}
