% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions_custom.R
\name{adjust_predictions_custom}
\alias{adjust_predictions_custom}
\title{Change or add variables}
\usage{
adjust_predictions_custom(x, ..., .pkgs = character(0))
}
\arguments{
\item{x}{A \code{\link[=container]{container()}}.}

\item{...}{Name-value pairs of expressions. See \code{\link[dplyr:mutate]{dplyr::mutate()}}.}

\item{.pkgs}{A character string of extra packages that are needed to execute
the commands.}
}
\description{
Change or add variables
}
\examples{
library(dplyr)
library(modeldata)

post_obj <-
  container(mode = "classification") \%>\%
  adjust_equivocal_zone() \%>\%
  adjust_predictions_custom(linear_predictor = binomial()$linkfun(Class2))


post_res <- fit(
  post_obj,
  two_class_example,
  outcome = c(truth),
  estimate = c(predicted),
  probabilities = c(Class1, Class2)
)

predict(post_res, two_class_example)
}
