% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_calibration.R
\name{adjust_probability_calibration}
\alias{adjust_probability_calibration}
\title{Re-calibrate classification probability predictions}
\usage{
adjust_probability_calibration(x, calibrator)
}
\arguments{
\item{x}{A \code{\link[=container]{container()}}.}

\item{calibrator}{A pre-trained calibration method from the \pkg{probably}
package, such as \code{\link[probably:cal_estimate_logistic]{probably::cal_estimate_logistic()}}.}
}
\description{
Re-calibrate classification probability predictions
}
