% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_methods.R
\name{categorical_labeling}
\alias{categorical_labeling}
\title{categorical_labelling}
\usage{
categorical_labeling(categorical_clusters, defs, params)
}
\arguments{
\item{categorical_clusters}{Information about categorical clusters obtained from a tailor
object.}

\item{defs}{A matrix or data frame giving definitions of major phenotypes.}

\item{params}{A list of markers to use; must be subset of colnames(data).}
}
\value{
WRITE THIS.
}
\description{
If major phenotype definitions are available (e.g. "CD4 Naive"),
based on only a few of the markers
in the panel, label each categorical cluster by one of these major phenotypes.
}
