% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_methods.R
\name{categorical_labeling}
\alias{categorical_labeling}
\title{categorical_labelling}
\usage{
categorical_labeling(tailor_obj, defs)
}
\arguments{
\item{tailor_obj}{A tailor object, as obtained from tailor.learn.}

\item{defs}{A matrix or data frame giving definitions of major phenotypes.}
}
\value{
The tailor object, with updated information about categorical labels.
}
\description{
If major phenotype definitions are available (e.g. "CD4 Naive"),
based on only a few of the markers
in the panel, label each categorical cluster by one of these major phenotypes.
}
