% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_methods.R
\name{customize_1D_mixtures}
\alias{customize_1D_mixtures}
\title{customize_1D_mixtures}
\usage{
customize_1D_mixtures(
  data,
  to_customize,
  mixtures_1D,
  sample_fraction = 0.2,
  seed = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{A flowFrame or a matrix containing events along the rows, markers along columns.}

\item{to_customize}{A named list, whose names are markers, and values are the number of mixture
components to learn for each marker.}

\item{mixtures_1D}{1D mixture models, obtained from get_1D_mixtures.}

\item{sample_fraction}{A number between 0 and 1: the fraction of data points used in the
calculation of 1D mixture components, to improve runtime.}

\item{seed}{A seed for the random generator, used for subsampling and for the initialization
of the 1D mixture components. The results are qualitatively the same, irrespective of the
seed chosen.}

\item{verbose}{Boolean flag; if true, outputs timing and milestone information.}
}
\value{
Updated version of mixtures_1D.
}
\description{
After visual inspection of 1D mixtures, manually specify the number of
mixture components to learn for some of the markers.
}
