% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_adapter.R
\name{find_adapter}
\alias{find_adapter}
\title{Find adapter region}
\usage{
find_adapter(bps, pident_threshold = 0.4, start_threshold = 0.25, merge = TRUE)
}
\arguments{
\item{bps}{data.frame; a data frame of breakpoints}

\item{pident_threshold}{numeric; the lowest accepted percentage identity for
a region to be considered a "conserved" region.}

\item{start_threshold}{numeric; the highest accepted starting position for
conserved region to be considered an "N-terminal" region, expressed as the
ratio of the aligned sequence length.}

\item{merge}{logical; whether neighboring conserved regions should be merged.}
}
\value{
A data frame of class "adapter".
}
\description{
This functions takes a data frame of breakpoints from a pairwise global
alignment and detects conserved N-terminal regions, optinally merging
neighboring conserved regions.
}
\details{
The function returns a single row for each pairwise comparison. If
multiple conserved N-terminal regions are identified, the function returns
the first one. If no conserved N-terminal regions are identified, the
function returns a row with NA values for the start, end, mean_score, and
pident columns.
}
\examples{
data(rbps)
ps <- position_scores("MN395291-1", "ON513429-1", data = rbps)
bps <- find_breakpoints(ps, Nmax = 5)
find_adapter(bps)
}
