% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_all_adapters.R
\name{find_all_adapters}
\alias{find_all_adapters}
\title{Find all adapters for a set of RBPs}
\usage{
find_all_adapters(
  ids,
  data,
  id_var = "Core_ORF",
  seq_var = "translation",
  submat = "BLOSUM80",
  method = "cemean",
  pident_threshold = 0.4,
  start_threshold = 0.25,
  merge = TRUE,
  cores = 1,
  verbose = getOption("verbose"),
  ...
)
}
\arguments{
\item{ids}{character; the names of RBPs to compare.}

\item{data}{data.frame; a data frame that contains IDs and sequences}

\item{id_var}{character; variable within \code{data} that stores IDs.}

\item{seq_var}{character; variable within \code{data} that stores sequences.}

\item{submat}{character; the substitution matrix. See 
\code{position_scores()} for more information.}

\item{method}{character; the method to use for finding breakpoints. Either
"cemean", "plateau", or "window". See \code{find_breakpoints()} for more
information}

\item{pident_threshold}{numeric; the lowest accepted percentage identity for
a region to be considered a "conserved" region.}

\item{start_threshold}{numeric; the highest accepted starting position for
conserved region to be considered an "N-terminal" region, expressed as the
ratio of the aligned sequence length.}

\item{merge}{logical; whether neighboring conserved regions should be merged.}

\item{cores}{integer; number of CPU cores to use.}

\item{verbose}{logical; should verbose messages be printed to the console?}

\item{...}{additional arguments to pass to breakpoint detection. See
\code{find_breakpoints()} for more information.}
}
\value{
A data.frame of class "adapter".
}
\description{
Look at all pairs between a set of RBPs, align, detect breakpoints, find
adapters and combine them into a single data frame.
}
\examples{
data(rbps)
find_all_adapters(rbps$Core_ORF[1:3], data = rbps)
}
