% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_methods.R
\name{plot_tsne_bin_centers}
\alias{plot_tsne_bin_centers}
\title{plot_tsne_bin_centers}
\usage{
plot_tsne_bin_centers(tailor_obj)
}
\arguments{
\item{tailor_obj}{A tailor object, as obtained from tailor.learn.}
}
\value{
A 2d reduced plot of bin centers, which can be seen as representatives for the cell
population. Color coded by major phenotype.
}
\description{
Plot a t-SNE reduction to 2d of the cluster centroids.
}
\examples{
fileName <- system.file("extdata", "sampled_flowset_old.rda", package = "Tailor")
load(fileName)
tailor_params <- flowCore::colnames(fs_old)[c(7:9, 11:22)]
tailor_obj <- tailor_learn(data = fs_old,
                          params = tailor_params,
                          mixture_components = 50)
plot_tsne_bin_centers(tailor_obj)
}
