% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor_methods.R
\name{plot_tsne_clusters}
\alias{plot_tsne_clusters}
\title{plot_tsne_clusters}
\usage{
plot_tsne_clusters(tailor_obj, tailor_pred)
}
\arguments{
\item{tailor_obj}{A tailor object, as obtained from tailor.learn.}

\item{tailor_pred}{A list of cluster assignments, as obtained from tailor.predict.}
}
\value{
A 2d reduced plot of cluster centroid, color-coded by major phenotype.
}
\description{
Plot a t-SNE reduction to 2d of the cluster centroids.
}
\examples{
fileName <- system.file("extdata", "sampled_flowset_old.rda", package = "Tailor")
load(fileName)
tailor_params <- flowCore::colnames(fs_old)[c(7:9, 11:22)]
tailor_obj <- tailor_learn(data = fs_old,
                          params = tailor_params,
                          mixture_components = 50)
tailor_pred <- tailor_predict(fs_old, tailor_obj)

plot_tsne_clusters(tailor_obj, tailor_pred)
}
