% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tailor-internals}
\alias{tailor-internals}
\alias{tailor_fully_trained}
\title{Internal tailor functions}
\usage{
tailor_fully_trained(x)
}
\description{
Utilities for use in downstream packages.
}
\keyword{internal}
