% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor-package.R
\docType{package}
\name{tailor-package}
\alias{tailor-package}
\title{tailor: Iterative Steps for Postprocessing Model Predictions}
\description{
Postprocessors refine predictions outputted from machine learning models to improve predictive performance or better satisfy distributional limitations. This package introduces 'tailor' objects, which compose iterative adjustments to model predictions. A number of pre-written adjustments are provided with the package, like calibration and equivocal zones, as well as utilities to compose new ones. Tailors are tightly integrated with the 'tidymodels' framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/tailor}
  \item \url{https://tailor.tidymodels.org}
  \item Report bugs at \url{https://github.com/tidymodels/tailor/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co}

Authors:
\itemize{
  \item Simon Couch \email{simon.couch@posit.co}
  \item Hannah Frick \email{hannah@posit.co}
  \item Emil HvitFeldt \email{emil.hvitfeldt@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
