% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailor.R
\name{tailor}
\alias{tailor}
\title{Declare post-processing for model predictions}
\usage{
tailor(type = "unknown", outcome = NULL, estimate = NULL, probabilities = NULL)
}
\arguments{
\item{type}{The model sub-type. Possible values are \code{"unknown"}, \code{"regression"},
\code{"binary"}, or \code{"multiclass"}.}

\item{outcome}{The name of the outcome variable.}

\item{estimate}{The name of the point estimate (e.g. predicted class). In
tidymodels, this corresponds to column names \code{.pred}, \code{.pred_class}, or
\code{.pred_time}.}

\item{probabilities}{The names of class probability estimates (if any). For
classification, these should be given in the order of the factor levels of
the \code{estimate}.}
}
\description{
Declare post-processing for model predictions
}
\examples{

tailor()
}
